fix should revalidate functions to incorporate form data
