+++ client/src/routes/Quiz/Quiz.data.tsx
@@ -16,15 +16,15 @@ import {
 export const quizShouldRevalidate: ShouldRevalidateFunction = ({
     currentUrl,
     nextUrl,
-    defaultShouldRevalidate,
+    formData,
 }) => {
     const currentUrlParams = new URLSearchParams(currentUrl.search);
     const nextUrlParams = new URLSearchParams(nextUrl.search);
 
     return (
+        !!formData ||
         currentUrlParams.get("page") !== nextUrlParams.get("page") ||
-        currentUrlParams.get("per_page") !== nextUrlParams.get("per_page") ||
-        defaultShouldRevalidate
+        currentUrlParams.get("per_page") !== nextUrlParams.get("per_page")
     );
 };
